requirements.txt
requests==2.32.3
requests-cache==1.2.1
pandas==2.2.3
matplotlib==3.10.1
numpy==2.2.3
python-dotenv==1.0.1